requirements.txt
python-telegram-bot[job-queue]
feedparser
requests
pytest
tweepy
google-genai
python-dotenv
pillow
beautifulsoup4
google-search-results
flask
